move TypeDef definition to wasm-type.cpp
